move print buttons to components
